refactor: split query-engine.ts into six retrieval stages
Split the 1,614-line monolithic QueryEngine class into focused stage modules:

1. query-normalizer.ts — question normalization, ID/lexeme/route detection
2. candidate-seeder.ts — pattern/route/lexeme scoring, heuristic seeds,
   session context
3. candidate-ranker.ts — ranking, route-wins logic, initial node selection
4. frontier-expander.ts — bounded graph expansion with hop budget and
   role-coverage
5. answer-projector.ts — route/pattern/Part C answers, confidence, gaps
6. synthesis-adapter.ts — optional LLM synthesis handoff with deterministic
   fallback

QueryEngine is now an orchestrator that wires stages together and exposes
query(), trace(), inspect(), readDoc(), inspectAnchor(), expandCitations().

All 39 existing tests pass without modification.

Closes #2
